Add a safety valve to room storage
Harvesters will drop excess resources from storage if
it is full.
Under normal operation there should be a few 100k of spare space,
so if this valve triggers something is really going wrong.

This will protect the room from situations where there is no more
space for energy in the storage.