Let upgraders choose source differently
Previously they always chose the closest source,
now they take the one farther from the primary spawn.

In a good room this is the same (I used to build the spawn
far away from the controller source). But if not, there is now
no conflict between harvesters/builders and upgraders.