Add global class AbsolutePosition
This is capable to be serialized/deserialized
into memory, so it can be used instead of flags to tell
creeps where to go (e.g. in operations).

Additionally it plugs into the multi-shard navigation
offered by goTo, so by using this kind of position
it should be possible to effectively tell creeps
to move to a different shard.