Make defenders more passable
Defenders are often standing in the way for other
creeps, since they are very close to the exit and
do not announce that they are stopped.

This also breaks blocked-detection pretty effectively.

As a workaround they will now announce their stopped state,
so that other creeps route around them and also move further away from exits.