Stop remote mining when there is too much energy
There is no point in overfarming, if we can't
get rid of excess energy.

Usually we have plenty of opportunities to get rid of
energy:

* controller upgrades
* wall reinforcements
* power refining

If all of those opportunities are used (or we intentionally
skip one of them), we have to make sure that the storage in a room does not
overflow.