Build roads to stores
So far roads have been built to the resources
being harvested.
In the best case the road will go two pieces too far
through the store.
In the worst case the road will go a way that is way off
of the store, which means carriers will be unable to properly
pick up from the store.