Swap preferred sources of harvesters and upgraders
This is mostly relevant for early rooms.
Assuming that extensions/spawns are mostly centered between
sources anyways, so they are not too sensitive to distance.

However, regardless of upgrader preference, as soon as they have
carriers, these will be more efficient at draining a close source.
Thus a harvester SHOULD be able to get a larger yield from
the source further away from the controller.