Check off discovered rooms earlier
Previously a room was only removed from the
target list, if it was the current main target
of the discoverer.

Though path finding often caused the discoverer to
pass through rooms later on the list first.

Now all rooms visited by the discoverer are considered done,
regardless of their list position.